+++ workbench-client.code-workspace
@@ -65,6 +65,7 @@
       "gteq",
       "viewports",
       "clusterer",
+      "combinators",
     ],
     "cSpell.flagWords": [
       "yuo",
